+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+*Version 1.4.2* **August 6, 2023**
+- Fixed an issue with the new UIImage(color:) extension. It wasn't rendering at the correct size.
+- Added tests for the new UIImage extension into the test harness (the first tab actually tests it).
+
 *Version 1.4.2* **August 4, 2023**
 - Documentation fixes.